requirements.txt
fastapi~=0.115.0
pydantic~=2.9.2
starlette~=0.38.6
SQLAlchemy~=2.0.35
python-dotenv~=1.0.1
mysql-connector-python~=9.0.0